requirements.txt
Django==3.1
psycopg2-binary>=2.8
djangorestframework==3.11.1
gmplot==1.4.1